requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5
Pillow>=10.1.0
reportlab>=4.0.8
xlsxwriter>=3.1.9
python-dotenv>=1.0.0
django-mptt>=0.15.0  # لدعم شجرة الحسابات
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
django-widget-tweaks>=1.5.0
mysqlclient>=2.2.0
